add tests, move function to outputMonitor
