+++ src/vs/workbench/contrib/terminalContrib/chatAgentTools/browser/tools/monitoring/outputMonitor.ts
@@ -21,7 +21,6 @@ import { ChatAgentLocation } from '../../../../../chat/common/constants.js';
 import { ChatMessageRole, ILanguageModelsService } from '../../../../../chat/common/languageModels.js';
 import { IToolInvocationContext } from '../../../../../chat/common/languageModelToolsService.js';
 import { ITaskService } from '../../../../../tasks/common/taskService.js';
-import { detectsInputRequiredPattern } from '../../executeStrategy/executeStrategy.js';
 import { ILinkLocation } from '../../taskHelpers.js';
 import { IConfirmationPrompt, IExecution, IPollingResult, OutputMonitorState, PollingConsts } from './types.js';
 import { getTextResponseFromStream } from './utils.js';
@@ -711,4 +710,42 @@ type SuggestedOption = string | { description: string; option: string };
 interface ISuggestedOptionResult {
 	suggestedOption?: SuggestedOption;
 	sentToTerminal?: boolean;
+}
+
+
+
+// PowerShell-style multi-option line (supports [?] Help and optional default suffix) ending in whitespace
+const PS_CONFIRM_RE = /\s*(?:\[[^\]]\]\s+[^\[]+\s*)+(?:\(default is\s+"[^"]+"\):)?\s+$/;
+
+// Bracketed/parenthesized yes/no pairs at end of line: (y/n), [Y/n], (yes/no), [no/yes]
+const YN_PAIRED_RE = /(?:\(|\[)\s*(?:y(?:es)?\s*\/\s*n(?:o)?|n(?:o)?\s*\/\s*y(?:es)?)\s*(?:\]|\))\s+$/i;
+
+// Same as YN_PAIRED_RE but allows a preceding '?' or ':' and optional wrappers e.g. "Continue? (y/n)" or "Overwrite: [yes/no]"
+const YN_AFTER_PUNCT_RE = /[?:]\s*(?:\(|\[)?\s*y(?:es)?\s*\/\s*n(?:o)?\s*(?:\]|\))?\s+$/i;
+
+// Confirmation prompts ending with (y) e.g. "Ok to proceed? (y)"
+const CONFIRM_Y_RE = /\(y\)\s*$/i;
+
+const LINE_ENDS_WITH_COLON_RE = /:\s*$/;
+
+const END = /\(END\)$/;
+
+const PASSWORD = /password[:]?$/i;
+
+const QUESTION = /\?[\(\)\s]*$/i;
+
+const PRESS_ANY_KEY_RE = /press a(?:ny)? key/i;
+
+export function detectsInputRequiredPattern(cursorLine: string): boolean {
+	return (
+		PS_CONFIRM_RE.test(cursorLine) ||
+		YN_PAIRED_RE.test(cursorLine) ||
+		YN_AFTER_PUNCT_RE.test(cursorLine) ||
+		CONFIRM_Y_RE.test(cursorLine) ||
+		LINE_ENDS_WITH_COLON_RE.test(cursorLine.trim()) ||
+		END.test(cursorLine) ||
+		PASSWORD.test(cursorLine) ||
+		QUESTION.test(cursorLine) ||
+		PRESS_ANY_KEY_RE.test(cursorLine)
+	);
 }
